Data tree expand/collapse animations off to reduce flicker when items are closed.
Setting animated to False because closing an RTI (instead of simply collapsing it) with
animation on is very ugly. Closing entails removing children from the tree. In that case
all the items below the closed node are then redrawn, which results in a lot of flicker.

Also fixed some comments and log messages.