Removed PandasPanelRti, Don't include volitale data in test data.
Pandas Panels have been obsolete since Pandas 0.25.0, and never been directly used in Argos.

Some test data was added directly to the MappingRti instead to the underlying dictionary. However, when you open and close the RTI in the tree, this data is lost. Therefore, we turn it off for now and perhaps solve this problem later.